pyproject.toml
[tool.poetry]
name = "langchain-er"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.2.12"
langserve = "^0.2.2"
fastapi = "^0.112.0"
langchain-openai = "^0.1.20"
python-dotenv = "^1.0.1"
uvicorn = "^0.30.5"
sse-starlette = "^2.1.3"
langfuse = "^2.43.1"
langgraph = "^0.2.34"
swarm = {git = "https://github.com/openai/swarm.git"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
